ai-service/requirements.txt
Flask==3.0.0
requests==2.31.0
flask-limiter==4.1.1
flask-talisman==1.1.0
flask==3.0.3
groq==0.9.0
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
